+++ CHANGELOG.md
@@ -38,7 +38,7 @@ _ `_optional` subpackage for managing optional dependencies
 - `comp_rep_columns` property for `Parameter`, `SearchSpace`, `SubspaceDiscrete`
   and `SubspaceContinuous` classes
 - Reworked mechanisms for surrogate input/output scaling configurable per class
-- `ParameterScalerProtocol` class for enabling user-defined input scaling mechanisms
+- `SurrogateProtocol` as an interface for user-defined surrogate architectures 
 
 ### Changed
 - Passing an `Objective` to `Campaign` is now optional
@@ -72,7 +72,8 @@ _ `_optional` subpackage for managing optional dependencies
 - Passing a dataframe via the `data` argument to the `transform` methods of
   `SearchSpace`, `SubspaceDiscrete` and `SubspaceContinuous` is no longer possible.
   The dataframe must now be passed as positional argument.
-- A deprecation error is thrown when attempting to use `register_custom_architecture`
+- The role of `register_custom_architecture` has been taken over by
+  `baybe.surrogates.base.SurrogateProtocol`
 
 ## [0.9.1] - 2024-06-04
 ### Changed